ExCLS behaviour (excls v0.1.0)

Elixir Command Line Shell.

A simple framework for writing line-oriented command interpreters.

Types

state()

@type state() :: term()

Behaviour state.

t()

@type t() :: module()

Behaviour module.

Callbacks

autocomplete(line, cursor_at, state)
(optional)

@callback autocomplete(line :: String.t(), cursor_at :: non_neg_integer(), state()) ::
  {:suggest, completed_line :: IO.chardata(), cursor_idx :: non_neg_integer(),
   options :: [IO.chardata()], next :: state()}
  | {:done, completed_line :: nil | IO.chardata(), next :: state()}

Autocomplete function.

Suggest autocomplete options when a tab key is pressed.

command(line, state)

@callback command(line :: String.t(), state()) ::
  {:ok, IO.chardata(), next :: state()}
  | {:stop, IO.chardata(), exit_code :: non_neg_integer()}

Command function.

Interpret the argument as though it had been typed in response to the prompt.

init()

@callback init() :: state()

Initialization function.

intro()
(optional)

@callback intro() :: IO.chardata()

Intro banner.

A string to issue as an intro or banner.

prompt(state)

@callback prompt(state()) :: IO.chardata()

Prompt function.

The prompt issued to solicit input.

Functions

run(behaviour, opts)

See ExCLS.Shell.run/2.