ExCLS.TTY.Ctrl (excls v0.1.0)

Ctrl bindings.

Functions

a()
(macro)

CTRL+a moves the cursor to the beginning of the line

d()
(macro)

CTRL+d delete one character to the right of the cursor

e()
(macro)

CTRL+e moves the cursor to the end of the line

left()
(macro)

CTRL+← move the cursor one word left

right()
(macro)

CTRL+→ move the cursor one word right